Issue #49. User can now select multiple variant files to delete

+++ src/gui/VariantMenuPanel.java
@@ -16,15 +16,12 @@ import javax.swing.event.ListSelectionEvent;
 import javax.swing.event.ListSelectionListener;
 import utility.FileUtility;
 
-public class VariantMenuPanel extends JPanel
-{
-	public VariantMenuPanel()
-	{
+public class VariantMenuPanel extends JPanel {
+	public VariantMenuPanel() {
 		initGuiComponents();
 	}
 
-	private void initGuiComponents()
-	{
+	private void initGuiComponents() {
 		setLayout(new GridBagLayout());
 		GridBagConstraints constraints = new GridBagConstraints();
 
@@ -34,13 +31,10 @@ public class VariantMenuPanel extends JPanel
 		constraints.anchor = GridBagConstraints.CENTER;
 
 		JButton createNewVariantButton = new JButton("Create New Variant");
-		createNewVariantButton.addActionListener(new ActionListener()
-		{
+		createNewVariantButton.addActionListener(new ActionListener() {
 			@Override
-			public void actionPerformed(ActionEvent event)
-			{
-				if (mPopupFrame == null)
-				{
+			public void actionPerformed(ActionEvent event) {
+				if (mPopupFrame == null) {
 					Driver.getInstance().setPanel(new CustomSetupPanel());
 				}
 			}
@@ -58,8 +52,7 @@ public class VariantMenuPanel extends JPanel
 		final JList variantList = new JList();
 		DefaultListModel variantListModel = new DefaultListModel();
 		String[] variantArray = FileUtility.getVariantsFileArrayNoClassic();
-		for (int i = 0; i < variantArray.length; i++)
-		{
+		for (int i = 0; i < variantArray.length; i++) {
 			variantListModel.addElement(variantArray[i]);
 		}
 		variantList.setModel(variantListModel);
@@ -73,30 +66,32 @@ public class VariantMenuPanel extends JPanel
 		constraints.ipadx = 7;
 		final JButton editButton = new JButton("Edit");
 		editButton.setEnabled(false);
-		editButton.addActionListener(new ActionListener()
-		{
+		editButton.addActionListener(new ActionListener() {
 			@Override
-			public void actionPerformed(ActionEvent event)
-			{
-				
+			public void actionPerformed(ActionEvent event) {
+
 			}
 		});
 		buttonPanel.add(editButton, constraints);
 
 		final JButton deleteButton = new JButton("Delete");
 		deleteButton.setEnabled(false);
-		deleteButton.addActionListener(new ActionListener()
-		{
+		deleteButton.addActionListener(new ActionListener() {
 			@Override
-			public void actionPerformed(ActionEvent event)
-			{
-				File variantFile = new File(System.getProperty("user.dir") + "\\variants\\"
-						+ variantList.getSelectedValue().toString());
-				((DefaultListModel) (variantList.getModel())).removeElementAt(variantList.getSelectedIndex());
-				variantFile.delete();
-						
-				if (((DefaultListModel) (variantList.getModel())).size() == 1)
-				{
+			public void actionPerformed(ActionEvent event) {
+				DefaultListModel dlm = (DefaultListModel) variantList
+						.getModel();
+				
+				if (variantList.getSelectedIndices().length > 0) {
+					int[] selectedIndices = variantList.getSelectedIndices();
+					for (int i = selectedIndices.length - 1; i >= 0; i--) {				
+						File variantFile = FileUtility.getVariantsFile(dlm.get(selectedIndices[i]).toString());					
+						dlm.removeElementAt(selectedIndices[i]);
+						variantFile.delete();
+					}
+				}
+
+				if (dlm.size() == 1) {
 					editButton.setEnabled(false);
 					deleteButton.setEnabled(false);
 					editDeletePanel.setVisible(false);
@@ -112,24 +107,20 @@ public class VariantMenuPanel extends JPanel
 		constraints.gridy = 1;
 		add(editDeletePanel, constraints);
 
-		variantList.addListSelectionListener(new ListSelectionListener()
-		{
+		variantList.addListSelectionListener(new ListSelectionListener() {
 			@Override
-			public void valueChanged(ListSelectionEvent arg0)
-			{
+			public void valueChanged(ListSelectionEvent arg0) {
 				deleteButton.setEnabled(true);
-				//TODO: enable this button once editing variants works
-//				editButton.setEnabled(true);
+				// TODO: enable this button once editing variants works
+				// editButton.setEnabled(true);
 			}
 		});
 
 		JButton backButton = new JButton("Return to Main Menu");
 		backButton.setToolTipText("Return to the Main Menu");
-		backButton.addActionListener(new ActionListener()
-		{
+		backButton.addActionListener(new ActionListener() {
 			@Override
-			public void actionPerformed(ActionEvent event)
-			{
+			public void actionPerformed(ActionEvent event) {
 				Driver.getInstance().revertToMainPanel();
 			}
 		});
